# src/MarqSpec.TradingCopilot.Client/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Trading Co-Pilot API",
    "description": "The single-operator futures trading co-pilot BFF (R-10). Generated from the live minimal-API routes (gh#604) — the source of truth the README links to, replacing the hand-kept table.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/connections": {
      "post": {
        "tags": ["Connections"],
        "summary": "Create a firm connection (login).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      },
      "get": {
        "tags": ["Connections"],
        "summary": "List the operator's connections.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
    "/connections/{id}/accounts": {
      "get": {
        "tags": ["Connections"],
        "summary": "List the accounts discovered through a connection.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
    "/connections/{id}/accounts/discover": {
      "post": {
        "tags": ["Connections"],
        "summary": "Discover accounts through a connection (mode from the firm's declared conventions).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
    "/connections/{id}/credentials": {
      "put": {
        "tags": ["Connections"],
        "summary": "Rotates which credential key a connection names (gh#210, R-17, ADR-0015).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The connection.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The caller's cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateCredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated connection, or a refusal."
          }
        },
        "security": [{}]
      }
    },
    "/connections/{id}": {
      "delete": {
        "tags": ["Connections"],
        "summary": "Deactivates a connection and cascades to its accounts — a soft delete (gh#210, R-17, R-20).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The connection.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok on deactivation (idempotent), or 404 when it is not the operator's."
          }
        },
        "security": [{}]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateConnectionRequest": {
        "required": ["firmId", "platform", "credentialKey"],
        "type": "object",
        "properties": {
          "firmId": {
            "type": "string",
            "description": "The firm this login belongs to.",
            "format": "uuid"
          },
          "platform": {
            "type": "string",
            "description": "The platform key (e.g. `projectx`). Only wired adapters are accepted."
          },
          "credentialKey": {
            "type": "string",
            "description": "The <b>non-secret</b> env-entry name the credentials live under."
          }
        },
        "description": "The request to register a firm login (ADR-0016: one per firm × platform)."
      },
...
      "RotateCredentialsRequest": {
        "required": ["credentialKey"],
        "type": "object",
        "properties": {
          "credentialKey": {
            "type": "string",
            "description": "The new credential key this connection should name."
          }
        },
        "description": "Rotate which credential key a connection names (gh#210, ADR-0015)."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT issued by POST /auth/login, sent as: Authorization: Bearer {token}.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}